Add LookupInput component to Softphone components and update ControlPanel to use it
