+++ src/Softphone/components/index.ts
@@ -10,3 +10,4 @@ export { TimeIndicator } from "./TimeIndicator/TimeIndicator";
 export { Contact } from "./Contact/Contact";
 export { ContactListItem } from "./ContactListItem/ContactListItem";
 export { SideBarPanel } from "./SideBarPanel/SideBarPanel";
+export { LookupInput } from "./LookupInput/LookupInput";
